Take the golden images logic out of ssp.go
The golden images logic is meaningfully large, and takes a big part of
the ssp.go file. That creates confusions and make the file to large,
hard to read and maintained. In addition, many golden image unit tests
are duplicated, with several level of the flow.

This commit introduces the new golden-images package with all the golden images
(DataImportCron) logic, with more accurate unit test suite. Then it
removes all the golden images logic and tests from the handlers
packages, and use the new package instead.
