+++ controllers/handlers/golden-images/testFiles/dataImportCronTemplates/dataImportCronTemplates.yaml
