+++ controllers/handlers/golden-images/testFiles/dataImportCronTemplates/dataImportCronTemplatesNoAnnotation.yaml
