+++ controllers/handlers/golden-images/testFiles/dataImportCronTemplates/dataImportCronTemplatesWithImageStream.yaml
