+++ tests/func-tests/golden_image_test.go
@@ -25,7 +25,7 @@ import (
 	sspv1beta3 "kubevirt.io/ssp-operator/api/v1beta3"
 
 	hcov1beta1 "github.com/kubevirt/hyperconverged-cluster-operator/api/v1beta1"
-	"github.com/kubevirt/hyperconverged-cluster-operator/controllers/handlers"
+	goldenimages "github.com/kubevirt/hyperconverged-cluster-operator/controllers/handlers/golden-images"
 	tests "github.com/kubevirt/hyperconverged-cluster-operator/tests/func-tests"
 )
 
@@ -399,19 +399,19 @@ var _ = Describe("golden image test", Label("data-import-cron"), Serial, Ordered
 					hcoDict, exists := getHCODICT(hc, dict.Name)
 					g.Expect(exists).To(BeTrue(), "should have the DICT in the HCO status")
 
-					if _, hcoAnnotationExists := hcoDict.Annotations[handlers.MultiArchDICTAnnotation]; !hcoAnnotationExists {
+					if _, hcoAnnotationExists := hcoDict.Annotations[goldenimages.MultiArchDICTAnnotation]; !hcoAnnotationExists {
 						GinkgoLogr.Info(fmt.Sprintf("The %q DICT does not have the multi-arch annotation in the HCO status; skipping it", dict.Name))
 						continue
 					}
 
-					multiArchAnnotation, exists := dict.GetAnnotations()[handlers.MultiArchDICTAnnotation]
+					multiArchAnnotation, exists := dict.GetAnnotations()[goldenimages.MultiArchDICTAnnotation]
 
 					g.Expect(exists).To(BeTrue(), "should have the multi-arch annotation in the DICT")
 					g.Expect(multiArchAnnotation).ToNot(Equal(""), "should have a value in the the multi-arch annotation; the %q DICT is:\n%#v", dict.Name, dict)
 
 					expectedArches := getExpectedArchs(hcoDict.Status.OriginalSupportedArchitectures, archs)
 
-					g.Expect(multiArchAnnotation).To(Equal(expectedArches), "the SSP %q DICT %q annotation should be %q", dict.Name, handlers.MultiArchDICTAnnotation, expectedArches)
+					g.Expect(multiArchAnnotation).To(Equal(expectedArches), "the SSP %q DICT %q annotation should be %q", dict.Name, goldenimages.MultiArchDICTAnnotation, expectedArches)
 				}
 
 			}).WithTimeout(10 * time.Second).
@@ -437,7 +437,7 @@ var _ = Describe("golden image test", Label("data-import-cron"), Serial, Ordered
 				}
 
 				customDictArchs := append(archs, "someOtherArch1", "someOtherArch2")
-				hcCustomDict.Annotations[handlers.MultiArchDICTAnnotation] = strings.Join(customDictArchs, ",")
+				hcCustomDict.Annotations[goldenimages.MultiArchDICTAnnotation] = strings.Join(customDictArchs, ",")
 
 				hc.Spec.DataImportCronTemplates = []hcov1beta1.DataImportCronTemplate{hcCustomDict}
 
@@ -462,14 +462,14 @@ var _ = Describe("golden image test", Label("data-import-cron"), Serial, Ordered
 				g.Expect(idx).To(BeNumerically(">", -1), "should have the custom-dict in the SSP")
 				sspDict := ssp.Spec.CommonTemplates.DataImportCronTemplates[idx]
 
-				multiArchAnnotation, exists := sspDict.GetAnnotations()[handlers.MultiArchDICTAnnotation]
+				multiArchAnnotation, exists := sspDict.GetAnnotations()[goldenimages.MultiArchDICTAnnotation]
 
 				g.Expect(exists).To(BeTrue(), "should have the multi-arch annotation in the DICT")
 				g.Expect(multiArchAnnotation).ToNot(BeEmpty(), "should have a value in the the multi-arch annotation")
 
-				expectedArches := getExpectedArchs(hcCustomDict.Annotations[handlers.MultiArchDICTAnnotation], archs)
+				expectedArches := getExpectedArchs(hcCustomDict.Annotations[goldenimages.MultiArchDICTAnnotation], archs)
 
-				g.Expect(multiArchAnnotation).To(Equal(expectedArches), "the SSP %q DICT %q annotation should be %q", "custom-dict", handlers.MultiArchDICTAnnotation, expectedArches)
+				g.Expect(multiArchAnnotation).To(Equal(expectedArches), "the SSP %q DICT %q annotation should be %q", "custom-dict", goldenimages.MultiArchDICTAnnotation, expectedArches)
 
 			}).WithTimeout(10 * time.Second).
 				WithPolling(500 * time.Millisecond).
@@ -496,7 +496,7 @@ var _ = Describe("golden image test", Label("data-import-cron"), Serial, Ordered
 				}
 
 				customDictArchs := append(archs, "someOtherArch1", "someOtherArch2")
-				hcCustomDict.Annotations[handlers.MultiArchDICTAnnotation] = strings.Join(customDictArchs, ",")
+				hcCustomDict.Annotations[goldenimages.MultiArchDICTAnnotation] = strings.Join(customDictArchs, ",")
 
 				hc.Spec.DataImportCronTemplates = []hcov1beta1.DataImportCronTemplate{hcCustomDict}
 
@@ -521,14 +521,14 @@ var _ = Describe("golden image test", Label("data-import-cron"), Serial, Ordered
 				g.Expect(idx).To(BeNumerically(">", -1), "should have the %q in the SSP", hcCustomDict.Name)
 				sspDict := ssp.Spec.CommonTemplates.DataImportCronTemplates[idx]
 
-				multiArchAnnotation, exists := sspDict.GetAnnotations()[handlers.MultiArchDICTAnnotation]
+				multiArchAnnotation, exists := sspDict.GetAnnotations()[goldenimages.MultiArchDICTAnnotation]
 
 				g.Expect(exists).To(BeTrue(), "should have the multi-arch annotation in the DICT")
 				g.Expect(multiArchAnnotation).ToNot(BeEmpty(), "should have a value in the the multi-arch annotation")
 
-				expectedArches := getExpectedArchs(hcCustomDict.Annotations[handlers.MultiArchDICTAnnotation], archs)
+				expectedArches := getExpectedArchs(hcCustomDict.Annotations[goldenimages.MultiArchDICTAnnotation], archs)
 
-				g.Expect(multiArchAnnotation).To(Equal(expectedArches), "the SSP %q DICT %q annotation should be %q", hcCustomDict.Name, handlers.MultiArchDICTAnnotation, expectedArches)
+				g.Expect(multiArchAnnotation).To(Equal(expectedArches), "the SSP %q DICT %q annotation should be %q", hcCustomDict.Name, goldenimages.MultiArchDICTAnnotation, expectedArches)
 
 				hc := tests.GetHCO(ctx, cli)
 				idx = slices.IndexFunc(hc.Status.DataImportCronTemplates, func(d hcov1beta1.DataImportCronTemplateStatus) bool {
@@ -537,7 +537,7 @@ var _ = Describe("golden image test", Label("data-import-cron"), Serial, Ordered
 				g.Expect(idx).To(BeNumerically(">", -1), "should have the %q in the HC status", hcCustomDict.Name)
 
 				hcoDictStatus := hc.Status.DataImportCronTemplates[idx]
-				g.Expect(hcoDictStatus.Annotations).To(HaveKeyWithValue(handlers.MultiArchDICTAnnotation, expectedArches))
+				g.Expect(hcoDictStatus.Annotations).To(HaveKeyWithValue(goldenimages.MultiArchDICTAnnotation, expectedArches))
 				g.Expect(hcoDictStatus.Status.OriginalSupportedArchitectures).To(Equal(originalSupportedArchitectures))
 				g.Expect(hcoDictStatus.Status.Conditions).To(BeEmpty())
 
@@ -560,7 +560,7 @@ var _ = Describe("golden image test", Label("data-import-cron"), Serial, Ordered
 
 					hcCustomDict.Name = "custom-dict"
 					hcCustomDict.Spec.ManagedDataSource = "custom-source"
-					delete(hcCustomDict.Annotations, handlers.MultiArchDICTAnnotation)
+					delete(hcCustomDict.Annotations, goldenimages.MultiArchDICTAnnotation)
 
 					hc.Spec.DataImportCronTemplates = []hcov1beta1.DataImportCronTemplate{hcCustomDict}
 
@@ -585,7 +585,7 @@ var _ = Describe("golden image test", Label("data-import-cron"), Serial, Ordered
 					g.Expect(idx).To(BeNumerically(">", -1), "should have the custom-dict DICT in the SSP")
 					sspDict := ssp.Spec.CommonTemplates.DataImportCronTemplates[idx]
 
-					g.Expect(sspDict.GetAnnotations()).ToNot(HaveKey(handlers.MultiArchDICTAnnotation))
+					g.Expect(sspDict.GetAnnotations()).ToNot(HaveKey(goldenimages.MultiArchDICTAnnotation))
 
 					hc := tests.GetHCO(ctx, cli)
 					idx = slices.IndexFunc(hc.Status.DataImportCronTemplates, func(d hcov1beta1.DataImportCronTemplateStatus) bool {
@@ -594,7 +594,7 @@ var _ = Describe("golden image test", Label("data-import-cron"), Serial, Ordered
 					g.Expect(idx).To(BeNumerically(">", -1), "should have the custom-dict in the HC status")
 
 					hcoDictStatus := hc.Status.DataImportCronTemplates[idx]
-					g.Expect(hcoDictStatus.Annotations).ToNot(HaveKey(handlers.MultiArchDICTAnnotation))
+					g.Expect(hcoDictStatus.Annotations).ToNot(HaveKey(goldenimages.MultiArchDICTAnnotation))
 					g.Expect(hcoDictStatus.Status.OriginalSupportedArchitectures).To(Equal(""))
 					g.Expect(hcoDictStatus.Status.Conditions).To(BeEmpty())
 
@@ -622,7 +622,7 @@ var _ = Describe("golden image test", Label("data-import-cron"), Serial, Ordered
 						hcCustomDict.Annotations = make(map[string]string)
 					}
 
-					hcCustomDict.Annotations[handlers.MultiArchDICTAnnotation] = "someOtherArch1,someOtherArch2"
+					hcCustomDict.Annotations[goldenimages.MultiArchDICTAnnotation] = "someOtherArch1,someOtherArch2"
 
 					hc.Spec.DataImportCronTemplates = []hcov1beta1.DataImportCronTemplate{hcCustomDict}
 
@@ -653,7 +653,7 @@ var _ = Describe("golden image test", Label("data-import-cron"), Serial, Ordered
 					g.Expect(idx).To(BeNumerically(">", -1), "should have the custom-dict in the HC status")
 
 					hcoDictStatus := hc.Status.DataImportCronTemplates[idx]
-					g.Expect(hcoDictStatus.Annotations).To(HaveKeyWithValue(handlers.MultiArchDICTAnnotation, ""))
+					g.Expect(hcoDictStatus.Annotations).To(HaveKeyWithValue(goldenimages.MultiArchDICTAnnotation, ""))
 					g.Expect(hcoDictStatus.Status.OriginalSupportedArchitectures).To(Equal("someOtherArch1,someOtherArch2"))
 
 					g.Expect(hcoDictStatus.Status.Conditions).To(HaveLen(1), "should have one condition in the DICT status")
@@ -685,7 +685,7 @@ var _ = Describe("golden image test", Label("data-import-cron"), Serial, Ordered
 					nextDICT := hc.Status.DataImportCronTemplates[1].DataImportCronTemplate
 					hcCustomDict.Spec.Template.Spec.Source.Registry = nextDICT.Spec.Template.Spec.Source.Registry.DeepCopy()
 
-					hcCustomDict.Annotations[handlers.MultiArchDICTAnnotation] = "someOtherArch1,someOtherArch2"
+					hcCustomDict.Annotations[goldenimages.MultiArchDICTAnnotation] = "someOtherArch1,someOtherArch2"
 
 					hc.Spec.DataImportCronTemplates = []hcov1beta1.DataImportCronTemplate{hcCustomDict}
 
@@ -716,7 +716,7 @@ var _ = Describe("golden image test", Label("data-import-cron"), Serial, Ordered
 					g.Expect(idx).To(BeNumerically(">", -1), "should have the %q in the HC status", hcCustomDict.Name)
 
 					hcoDictStatus := hc.Status.DataImportCronTemplates[idx]
-					g.Expect(hcoDictStatus.Annotations).To(HaveKeyWithValue(handlers.MultiArchDICTAnnotation, ""))
+					g.Expect(hcoDictStatus.Annotations).To(HaveKeyWithValue(goldenimages.MultiArchDICTAnnotation, ""))
 					g.Expect(hcoDictStatus.Status.OriginalSupportedArchitectures).To(Equal("someOtherArch1,someOtherArch2"))
 
 					g.Expect(hcoDictStatus.Status.Conditions).To(HaveLen(1), "should have one condition in the DICT status")
@@ -746,7 +746,7 @@ var _ = Describe("golden image test", Label("data-import-cron"), Serial, Ordered
 					}
 
 					originalSupportedArchitectures = hc.Status.DataImportCronTemplates[0].Status.OriginalSupportedArchitectures
-					hcCustomDict.Annotations[handlers.MultiArchDICTAnnotation] = "someOtherArch1,someOtherArch2"
+					hcCustomDict.Annotations[goldenimages.MultiArchDICTAnnotation] = "someOtherArch1,someOtherArch2"
 
 					hc.Spec.DataImportCronTemplates = []hcov1beta1.DataImportCronTemplate{hcCustomDict}
 
@@ -779,7 +779,7 @@ var _ = Describe("golden image test", Label("data-import-cron"), Serial, Ordered
 					By("check the we use the original annotation, not the custom one")
 					hcoDictStatus := hc.Status.DataImportCronTemplates[idx]
 					expectedAnnotation := getExpectedArchs(originalSupportedArchitectures, archs)
-					g.Expect(hcoDictStatus.Annotations).To(HaveKeyWithValue(handlers.MultiArchDICTAnnotation, expectedAnnotation))
+					g.Expect(hcoDictStatus.Annotations).To(HaveKeyWithValue(goldenimages.MultiArchDICTAnnotation, expectedAnnotation))
 					g.Expect(hcoDictStatus.Status.OriginalSupportedArchitectures).To(Equal(originalSupportedArchitectures))
 
 					g.Expect(hcoDictStatus.Status.Conditions).To(BeEmpty(), "should have no conditions in the DICT status")
